requirements.txt
# LLM and LangChain
langchain==0.3.13
langchain-groq==0.2.1
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4

# Vector Database
chromadb==0.5.23
sentence-transformers==3.3.1

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.0

# Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0

# Data Processing
pandas==2.2.3
numpy==2.2.1
